requirements.txt
# Stroke Prediction 
# Install with: pip install -r requirements.txt

# Core ML
autogluon.tabular==1.5.0

# Data
pandas>=2.0.0
numpy>=1.26.0

# Evaluation & Visualisation
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
joblib>=1.3.0

# Notebook (optional – only needed for EDA.ipynb)
jupyter>=1.0.0
ipykernel>=6.0.0
